requirements.txt
google-genai==1.68.0
psycopg2-binary==2.9.11
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.2
